Made functionality and routes for login and registering -> missing logout
made functionality to create and list users -> should now be able to create and list users. added email regex to validate user email. added cors to server and paths for users and auth. implemented update and delete methods for article. added method which enables us to send a token to other parts of app